Split storage low classes from holders.py
